move useTariffZoneFromLocation
